Fix build on NetBSD.
